Validações - Controle de Fluxo

+++ app/Http/Controllers/ScheduleController.php
@@ -15,6 +15,7 @@ class ScheduleController extends Controller
     public function index()
     {
         $schedule = Schedule::all();
+        
         return response()->json($schedule, 200);
     }   
 
@@ -26,8 +27,8 @@ class ScheduleController extends Controller
      */
     public function store(Request $request)
     {
-        //adicionar validacao
         $schedule = Schedule::create($request->all());
+
         return response()->json($schedule, 201);
     }
 
@@ -40,7 +41,12 @@ class ScheduleController extends Controller
     public function show($id)
     {
         $schedule = Schedule::find($id);
-        return response($schedule, 200);
+
+        if($schedule === null){
+            return response()->json('Recurso não encontrado!', 200);
+        }
+
+        return response()->json($schedule, 200);
     }
 
     /**
@@ -53,6 +59,11 @@ class ScheduleController extends Controller
     public function update(Request $request, $id)
     {
         $schedule = Schedule::find($id);
+
+        if($schedule === null){
+            return response()->json(['erro' => 'O recurso solicitado não existe!']);
+        }
+
         $schedule->fill($request->all());
         $schedule->save();
         return response()->json($schedule, 201);
@@ -67,9 +78,13 @@ class ScheduleController extends Controller
      */
     public function destroy($id)
     {
-        $schedule = Schedule::findOrFail($id);
+        $schedule = Schedule::find($id);
+
+        if($schedule === null){
+            return response()->json(['erro' => 'O recurso solicitado não existe!']);
+        }
+
         $schedule->delete();
         return response()->json('Agendamento cancelado!', 200);
-
     }
 }
